requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.26.0
opencv-python==4.8.1.78
pytesseract==0.3.10
pillow==10.0.1
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.0